Commiting changes for frontend routes and saving clue data.
