+++ client/src/components/ClueDataPage.css
@@ -1,0 +1,62 @@
+form {
+    width: 40%;
+    margin: 20px auto 0;
+    background-color: #f2f2f2;
+    padding: 20px;
+    border-radius: 10px;
+  }
+  .form-group {
+    display: flex;
+    align-items: center;
+    margin-bottom: 20px;
+  }
+
+  label, input[type="text"], input[type="checkbox"] {
+    display: inline-block;
+    vertical-align: middle;
+    text-align: left;
+  }
+  .required{
+    color: red;
+  }
+  
+  label {
+    width: 30%;
+    margin-bottom: 10px;
+    font-weight: bold;
+  }
+  
+  input[type="text"],
+  textarea {
+    width: 70%;
+    padding: 10px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    box-sizing: border-box;
+    margin-bottom: 20px;
+  }
+  input[type="checkbox"] {
+    display: inline-block;
+    width: 70%;
+    padding: 10px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    box-sizing: border-box;
+    margin-bottom: 20px;
+  }
+  
+  input[type="submit"] {
+    background-color: #B64F00;
+    color: white;
+    padding: 10px 20px;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    font-size: 16px;
+  }
+  
+  @media (max-width: 600px) {
+    form {
+      width: 80%;
+    }
+  }
