+++ client/src/components/NotFoundPage.js
@@ -1,0 +1,9 @@
+import React, { useState } from 'react';
+const NotfoundPage = () => {
+    return(
+        <div className="not-found-page">
+            <h1>Page does not exist !!</h1>
+        </div>
+    )
+}
+export default NotfoundPage;
